requirements.txt
# ==============================================================================
# IMPORTANT: This project requires Python 3.9, 3.10, or 3.11
# Apache Airflow 2.7.x does NOT support Python 3.12+ or 3.13+
#
# If you have Python 3.13 installed, please:
# 1. Install Python 3.11 from https://www.python.org/downloads/
# 2. Create venv with: py -3.11 -m venv .venv
# 3. Activate and install: .venv\Scripts\Activate.ps1; pip install -r requirements.txt
# ==============================================================================

# Core workflow orchestration (requires Python <3.12)
apache-airflow==2.7.3
apache-airflow-providers-amazon==8.13.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-http==4.7.0

# Data transformation
sqlalchemy==1.4.53  # Compatible with Airflow 2.7
dbt-core==1.6.14
dbt-postgres==1.6.14
dbt-snowflake==1.6.2

# Cloud + warehouse connectivity
snowflake-connector-python==3.5.0
boto3==1.34.0
botocore==1.34.0
psycopg2-binary==2.9.9

# Data generation
faker==22.0.0
mimesis==12.1.0

# Data quality & testing
great-expectations==0.18.8
pytest==7.4.4
pytest-cov==4.1.0

# Data manipulation & utilities
pandas==2.0.3  # Compatible with Python 3.11 and Airflow
numpy==1.24.4
python-dateutil==2.8.2
pytz==2023.3

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# AWS CLI is optional for local development. It's recommended to install it separately
# to avoid dependency conflicts. If you need awscli, install it globally or in a separate
# virtualenv: `pip install awscli`.

# Pin colorama to a version compatible with all Windows-specific requirements
# Many packages (awscli, pytest, pylint, great-expectations, dbt-core) have
# conflicting constraints. colorama==0.4.5 is a compatible choice on Windows.
# colorama==0.4.4

# API requests
requests==2.31.0
# urllib3: pin to 1.x to satisfy dbt-core dependency (dbt-core 1.6.x requires urllib3~=1.0)
urllib3~=1.26.16

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
tabulate==0.9.0

# Code quality & formatting
black==23.12.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3
mypy==1.8.0

# Testing
pytest-mock==3.12.0
pytest==7.4.4
pytest-cov==4.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Development tools
ipython==8.19.0
jupyterlab==4.0.10
jupyter==1.0.0
